Remove Type Casts

+++ source/BreachModel.cpp
@@ -3,10 +3,6 @@
 /**
  * Initializes a new breach with the given angle and max health
  *
- * This is an initializer.  It, combined with the constructor, produces the static
- * constructor create().  The initializer and normal constructor are private while
- * the static constructor is not.
- *
  * @param  a  The angle at which the breach exists
  * @param  health  The initial health of the breach
  * @param  p  the player id
@@ -14,7 +10,7 @@
  *
  * @return  true if the obstacle is initialized properly, false otherwise.
  */
-bool BreachModel::init(float a, int health, int p, float time) {
+bool BreachModel::init(float a, uint8_t health, uint8_t p, float time) {
 	angle = a;
 	this->health = health;
 	player = p;
@@ -24,10 +20,4 @@ bool BreachModel::init(float a, int health, int p, float time) {
 	return true;
 }
 
-/**
- * Disposes all resources and assets of this breach.
- *
- * Any assets owned by this object will be immediately released.  Once
- * disposed, a breach may not be used until it is initialized again.
- */
 void BreachModel::dispose() {}
